refactor: [UF-334] shared/ui의 컴포넌트의 배럴 익스포트 통일 및 경로 정리 #384
