+++ src/app/admin/posts/page.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import { Table } from '@/shared';
-import Header from '@/shared/ui/Header/Header';
-import Sidebar from '@/shared/ui/Sidebar/Sidebar';
+import { Header, Sidebar } from '@/shared';
 
 export default function AdminPostsPage() {
   const postsData = [
