+++ src/app/admin/posts/reported/page.tsx
@@ -5,13 +5,17 @@ import React, { useEffect, useState } from 'react';
 import { statisticsService } from '@/api/services/admin/statistics';
 import type { ReportsStatisticsData } from '@/api/types';
 import { useReportedPosts } from '@/features/admin/hooks/useReportedPosts';
+import {
+  Button,
+  Header,
+  Sidebar,
+  Skeleton,
+  AdminTable,
+  TableColumn,
+  TableActions,
+  BaseTableRow,
+} from '@/shared';
 import { useModal } from '@/shared/hooks/useModal';
-import { Button } from '@/shared/ui/Button/Button';
-import Header from '@/shared/ui/Header/Header';
-import Sidebar from '@/shared/ui/Sidebar/Sidebar';
-import { Skeleton } from '@/shared/ui/Skeleton';
-import { AdminTable } from '@/shared/ui/Table/AdminTable';
-import { TableColumn, TableActions, BaseTableRow } from '@/shared/ui/Table/Table.types';
 
 interface ReportedPostTableRow extends BaseTableRow {
   id: number;
