+++ src/app/admin/settings/page.tsx
@@ -1,5 +1,4 @@
-import Header from '@/shared/ui/Header/Header';
-import Sidebar from '@/shared/ui/Sidebar/Sidebar';
+import { Header, Sidebar } from '@/shared';
 
 export default function AdminSettingsPage() {
   return (
