+++ src/app/mypage/achievement/page.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react';
 import { achievementsAPI } from '@/api/services/mypage/achievement';
 import { AchievementBadge } from '@/features/mypage/components/AchievementBadge';
 import { Loading, Title } from '@/shared';
-import AchievementModal from '@/shared/ui/Modal/AchievementModal';
+import { AchievementModal } from '@/shared';
 import { useViewportStore } from '@/stores/useViewportStore';
 import { Achievement, SelectedAchievementState } from '@/types/Achievement';
