+++ src/app/mypage/notification/page.tsx
@@ -4,9 +4,9 @@ import { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 
 import { notificationAPI } from '@/api';
+import type { NotificationType } from '@/shared';
 import { Title } from '@/shared';
-import type { NotificationType } from '@/shared/ui/NotificationDropdown/NotificationDropdown.types';
-import { Switch } from '@/shared/ui/Switch/Switch';
+import { Switch } from '@/shared';
 
 const MypageNotificationPage = () => {
   const [notificationState, setNotificationState] = useState({
