+++ src/app/mypage/trade/page.tsx
@@ -13,7 +13,7 @@ import { TradeHistoryCard } from '@/features/mypage/components';
 import { TradeHistoryCardProps } from '@/features/mypage/types/TradeHistoryCard.types';
 import { useTradeHistory } from '@/hooks/useTradeHistory';
 import { BadgeState } from '@/shared';
-import { Button, Label, Tabs, TabsContent, TabsList, TabsTrigger } from '@/shared/ui';
+import { Button, Label, Tabs, TabsContent, TabsList, TabsTrigger } from '@/shared';
 import { useTradeTabStore } from '@/stores/useTradeTabStore';
 import { groupByDate } from '@/utils/groupByDate';
