+++ src/app/signal/page.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 
 import SignalTabContent from '@/features/signal/components/SignalTabContent';
 import { Title } from '@/shared';
-import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/shared/ui/Tabs';
+import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/shared';
 
 type TabType = 'orbit' | 'letters';
