+++ src/shared/layout/BottomNav/BottomNav.tsx
@@ -3,9 +3,9 @@
 import React from 'react';
 
 import { cn } from '@/lib/utils';
+import type { IconType } from '@/shared';
+import { Icon } from '@/shared';
 import { useNavigation } from '@/shared/hooks/useNavigation';
-import { Icon } from '@/shared/ui';
-import type { IconType } from '@/shared/ui/Icons/Icons.types';
 
 interface NavItem {
   id: string;
