+++ src/shared/ui/Modal/index.ts
@@ -1,5 +1,6 @@
 export { Modal } from './Modal';
 export { GlobalModal } from './GlobalModal';
+export { ReportedModal } from './ReportModal';
 export { modalVariants, modalOverlayVariants } from './modalVariants';
 export * from './Modal.types';
 export * from './AchievementModal';
